requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bcrypt==5.0.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.4
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.9
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.9
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
marshmallow==3.26.2
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
tenacity==9.1.4
Twisted==25.5.0
txaio==25.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
yarl==1.22.0
zope.interface==8.2
zstandard==0.25.0

# --- Core Framework & Servers ---
Django==6.0.2
gunicorn==23.0.0
daphne==4.2.1
psycopg2-binary==2.9.11
psycopg[binary]==3.2.4
dj-database-url==3.1.0
asgiref==3.11.1

# --- Auth & REST Framework ---
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0

# --- WebSockets & Redis ---
channels==4.3.2
channels_redis==4.3.0
redis==7.1.0

# --- AI & Cloud Integrations ---
google-cloud-vision==3.12.1
groq==1.0.0
langchain==1.2.9
langchain-community==0.4.1
cloudinary==1.44.1

# --- Utilities ---
requests==2.32.5
pydantic[email]==2.12.5
pydantic-settings==2.12.0
SQLAlchemy==2.0.46

dj-database-url
whitenoise
psycopg2-binary